refactor : chat service server code refactored
